+++ tests/collections/test_tomogram_reprocessing.py
@@ -1,22 +1,18 @@
 from sqlalchemy import select
 
-from pato.models.table import ProcessingJob, ProcessingJobParameter
+from pato.models.table import ProcessingJobParameter
 from pato.utils.database import db
 
 
 def test_post_user(mock_permissions, client):
     """Start reprocessing job for data collection"""
     resp = client.post(
-        "/dataCollections/6017406/tomograms/reprocessing",
+        "/dataCollections/6017406/reprocessing/tomograms",
         json={"pixelSize": 1, "tiltOffset": 1},
     )
     assert resp.status_code == 202
 
-    proc_id = db.session.scalars(
-        select(ProcessingJob.processingJobId)
-        .filter(ProcessingJob.dataCollectionId == 6017406)
-        .order_by(ProcessingJob.processingJobId.desc())
-    ).first()
+    proc_id = resp.json()["processingJobId"]
 
     # Should create five entries in the processing job parameters table with the
     # respective given values
@@ -36,16 +32,12 @@ def test_post_user(mock_permissions, client):
 def test_post_custom(mock_permissions, client):
     """Starts reprocessing job with custom parameters"""
     resp = client.post(
-        "/dataCollections/6017406/tomograms/reprocessing",
+        "/dataCollections/6017406/reprocessing/tomograms",
         json={"pixelSize": 51, "tiltOffset": 1},
     )
     assert resp.status_code == 202
 
-    proc_id = db.session.scalars(
-        select(ProcessingJob.processingJobId)
-        .filter(ProcessingJob.dataCollectionId == 6017406)
-        .order_by(ProcessingJob.processingJobId.desc())
-    ).first()
+    proc_id = resp.json()["processingJobId"]
 
     # Values added should match user provided values
 
@@ -63,16 +55,12 @@ def test_post_custom(mock_permissions, client):
 def test_post_message(mock_permissions, mock_pika, client):
     """Starts reprocessing job with custom parameters"""
     resp = client.post(
-        "/dataCollections/6017406/tomograms/reprocessing",
+        "/dataCollections/6017406/reprocessing/tomograms",
         json={"pixelSize": 51, "tiltOffset": 1},
     )
     assert resp.status_code == 202
 
-    proc_id = db.session.scalars(
-        select(ProcessingJob.processingJobId)
-        .filter(ProcessingJob.dataCollectionId == 6017406)
-        .order_by(ProcessingJob.processingJobId.desc())
-    ).first()
+    proc_id = resp.json()["processingJobId"]
 
     mock_pika.publish.assert_called_with(
         (
@@ -89,7 +77,7 @@ def test_post_message(mock_permissions, mock_pika, client):
 def test_no_tomogram(mock_permissions, client):
     """Try to process data collection with no tomogram"""
     resp = client.post(
-        "/dataCollections/993677/tomograms/reprocessing",
+        "/dataCollections/993677/reprocessing/tomograms",
         json={"pixelSize": 51, "tiltOffset": 1},
     )
 
@@ -99,7 +87,7 @@ def test_no_tomogram(mock_permissions, client):
 def test_tomogram_too_many(mock_permissions, client):
     """Try to process data collection with too many (3) tomograms already processed"""
     resp = client.post(
-        "/dataCollections/6017408/tomograms/reprocessing",
+        "/dataCollections/6017408/reprocessing/tomograms",
         json={"pixelSize": 51, "tiltOffset": 1},
     )
 
@@ -109,7 +97,7 @@ def test_tomogram_too_many(mock_permissions, client):
 def test_tomogram_no_tilt(mock_permissions, client):
     """Try to process data collection tomogram with no tilt alignment data"""
     resp = client.post(
-        "/dataCollections/6017409/tomograms/reprocessing",
+        "/dataCollections/6017409/reprocessing/tomograms",
         json={"pixelSize": 51, "tiltOffset": 1},
     )
 
@@ -119,7 +107,7 @@ def test_tomogram_no_tilt(mock_permissions, client):
 def test_tomogram_no_stack(mock_permissions, client):
     """Try to process data collection tomogram with no stack file"""
     resp = client.post(
-        "/dataCollections/6017409/tomograms/reprocessing",
+        "/dataCollections/6017409/reprocessing/tomograms",
         json={"pixelSize": 51, "tiltOffset": 1},
     )
 
@@ -129,7 +117,7 @@ def test_tomogram_no_stack(mock_permissions, client):
 def test_tomogram_invalid_micrograph(mock_permissions, client):
     """Try to process data collection tomogram with invalid micrograph name"""
     resp = client.post(
-        "/dataCollections/6017413/tomograms/reprocessing",
+        "/dataCollections/6017413/reprocessing/tomograms",
         json={"pixelSize": 51, "tiltOffset": 1},
     )
 
@@ -139,17 +127,13 @@ def test_tomogram_invalid_micrograph(mock_permissions, client):
 def test_tomogram_with_suffix(mock_permissions, client):
     """Process tomogram with cardinal suffix on stack file name"""
     resp = client.post(
-        "/dataCollections/6017406/tomograms/reprocessing",
+        "/dataCollections/6017406/reprocessing/tomograms",
         json={"pixelSize": 51, "tiltOffset": 1},
     )
 
     assert resp.status_code == 202
 
-    proc_id = db.session.scalars(
-        select(ProcessingJob.processingJobId)
-        .filter(ProcessingJob.dataCollectionId == 6017406)
-        .order_by(ProcessingJob.processingJobId.desc())
-    ).first()
+    proc_id = resp.json()["processingJobId"]
 
     assert (
         db.session.scalars(