+++ CHANGELOG.rst
@@ -3,6 +3,28 @@ Changelog
 ==========
 
 +++++++++
+v1.1.1 (12/07/2023)
++++++++++
+
+**Added**
+
+- Virtual host can be set in message queue configs
+
++++++++++
+v1.1.0 (11/07/2023)
++++++++++
+
+**Added**
+
+- GET endpoint for processing job parameters (:code:`dataCollections/{collectionId}/reprocessing`)
+- POST endpoint for firing off SPA reprocessing pipeline (:code:`dataCollections/{collectionId}/reprocessing/spa`)
+
+**Changed**
+
+- Processing job list endpoint is now ordered by both processing job and autoprocessing job
+- Tomogram reprocessing endpoint moved to :code:`dataCollections/{collectionId}/reprocessing/tomograms`
+
++++++++++
 v1.0.0 (20/06/2023)
 +++++++++